let car spawn at police man
